backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.4
python-multipart==0.0.6
gunicorn==21.2.0
celery==5.3.4
redis==4.6.0
